Moved src/{class,strict}.lua to src/lib/lua.
