+++ src/lib/lua/strict.lua
