+++ src/lib/protocol/icmp/header.lua
@@ -1,4 +1,3 @@
-require("class")
 local ffi = require("ffi")
 local C = ffi.C
 local header = require("lib.protocol.header")
