+++ src/lib/protocol/icmp/nd/header.lua
@@ -1,4 +1,3 @@
-require("class")
 local ffi = require("ffi")
 local proto_header = require("lib.protocol.header")
 local tlv = require("lib.protocol.icmp.nd.options.tlv")
