+++ src/lib/protocol/icmp/nd/options/lladdr.lua
@@ -1,4 +1,3 @@
-require("class")
 local ffi = require("ffi")
 
 local lladdr = subClass(nil, 'new_from_mem')
